server/requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mido==1.3.3
packaging==25.0
pygame==2.6.1
pyserial==3.5
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
